Add Dependency Install Conditions
Add conditions to the rest of the custom Python install scripts/modules.

+++ src/packages/04_nvm.py
@@ -1,2 +1,8 @@
 from subprocess import call
-call("curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash", shell=True)
+from pathlib import Path
+from src.os import is_installed
+
+nvm = Path.home() / ".nvm"
+
+if not nvm.exists() and not is_installed("nvm"):
+    call("curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash", shell=True)
